DentalVision/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
ultralytics==8.1.0
anthropic==0.18.1
Pillow>=10.0.0
python-dotenv==1.0.0
requests==2.31.0
numpy>=1.24.0
